+++ programs/uxd/src/mango_utils/mod.rs
@@ -1,9 +1,7 @@
-pub mod limit_utils;
 pub mod order_delta;
 pub mod perp_account_utils;
 pub mod perp_info;
 
-pub use limit_utils::*;
 pub use order_delta::*;
 pub use perp_account_utils::*;
 pub use perp_info::*;