+++ programs/uxd/src/test/mango_utils/mod.rs
@@ -1,3 +1,2 @@
-mod test_limit_utils;
 mod test_order_delta;
 mod test_perp_account_utils;